+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/TeleOpExample.java
@@ -7,7 +7,6 @@ import com.seattlesolvers.solverslib.gamepad.GamepadEx;
 import com.seattlesolvers.solverslib.gamepad.GamepadKeys;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.teamcode.config.TuningConfig;
 import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
 import org.firstinspires.ftc.teamcode.subsystems.Drivetrain;
 import org.firstinspires.ftc.teamcode.util.BulkReads;
@@ -55,13 +54,13 @@ public class TeleOpExample extends CommandOpMode {
 
         // Read -> process -> write (section 4, rule 2).
         double cap = driver.getButton(GamepadKeys.Button.LEFT_BUMPER)
-                ? TuningConfig.driveSlowModeCap
-                : TuningConfig.driveSpeedCap;
+                ? Drivetrain.driveSlowModeCap
+                : Drivetrain.driveSpeedCap;
 
         // Field-centric: Pedro rotates strafe/forward by the Pinpoint heading before applying power.
         // Sign convention from PedroTeleOpSample: (-leftY, -leftX, -rightX).
         // TODO: if a direction is backwards on the robot, flip that sign.
-        double dz = TuningConfig.driveDeadzone;
+        double dz = Drivetrain.driveDeadzone;
         double forward = applyDeadzone(-driver.getLeftY(), dz);
         double strafe  = applyDeadzone(-driver.getLeftX(), dz);
         double turn    = applyDeadzone(-driver.getRightX(), dz);